swiftvalidator/src/net/uglobal/swiftvalidator/swift/field/Field70F.java

package net.uglobal.swiftvalidator.swift.field;



/**
 * Field format : (qualifier)(narrative)
 * 
 * AUTO GENERATED - DO NOT HAND MODIFY
 * Generated on Sun Dec 09 17:22:28 EST 2012
 * 
 */
public class Field70F
    extends net.uglobal.swiftvalidator.swift.field.Field
{


    public Field70F(String validateData, net.uglobal.swiftvalidator.xml.beans.Field xmlField) {
        super(validateData, xmlField);
    }

    @Override
    protected void parse() {
        addComponents(net.uglobal.swiftvalidator.swift.field.FieldComponentParseUtil.parse(validateData, getContentFormat()));
    }

    public String getQualifier() {
        return getComponent1();
    }

    public String getNarrative() {
        return getComponent2();
    }

    public String toString() {
        return (new StringBuilder()).append("[" ).append( "qualifier=" + getComponent1()).append( ", narrative=" + getComponent2() ).append( "]").toString();
    }

}
